requirements.txt
asgiref
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.3
django-address==0.2.8
django-allauth==0.52.0
django-autoslug==1.9.9
django-cities==0.6.2
django-cities-light==3.10.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-crispy-forms==2.4
django-froala-editor==4.6.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-unfold==0.64.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
oauthlib==3.3.1
packaging
pillow
progressbar2
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
python-decouple
python-dotenv
python-utils
python3-openid
pytz
redis
requests
requests-oauthlib
service-identity
setuptools
six
sqlparse
stripe
swapper
tqdm
Twisted
txaio
typing_extensions
tzdata
Unidecode
urllib3
whitenoise
zope.interface
